# ARROW-11058: [Rust] [DataFusion] Implement coalesce batches operator