@Dandandan 2020-12-29
LGTM 👍  😎 very nice speedup, good to see it also works on the bigger batch sizes (so it's not only because of the "problematic" join). I think it is a nice building block which we can decide to use in optimizations.
We should still look at improving the join implementation, although this hides the problem quite a bit.