@andygrove 2020-12-29
> We should still look at improving the join implementation, although this hides the problem quite a bit.

https://issues.apache.org/jira/browse/ARROW-11030 is the tracking issue for the MutableArrayData performance issue